Use working dir if autoload.php is not found
`__DIR__` is the directory of the current file. If you run this file like that:
``` shell
composer require scrutinizer/ocular:~1.0.0
vendor/bin/ocular code-coverage:upload .....
```

It would not find the `autoload.php` file of Composer.
This is because Composer symlinks `bin/ocular` to `vendor/bin/ocular`.
This messes up with the path to `autoload.php`.

`getcwd` would use the current working directory from the shell.
Assuming you would most probably run `vendor/bin/ocular` this would work.

This is not breaking the previous workflow, but merely adding a fallback.